#3 refactor: saa 패턴 적용
- signin, signup, my Activity 제거
- mainActivity 선언
