+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.ANDANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".ui.signin.SignInActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ANDANDROID">
             <intent-filter>
@@ -21,14 +21,6 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".ui.signup.SignUpActivity"
-            android:exported="false">
-        </activity>
-        <activity
-            android:name=".ui.my.MyActivity"
-            android:exported="false">
         </activity>
     </application>
