+++ app/src/main/java/org/sopt/and/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.ui.my
+package org.sopt.and.ui.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,23 +6,13 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import org.sopt.and.core.designsystem.theme.ANDANDROIDTheme
 
-class MyActivity : ComponentActivity() {
+class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-
-        val userEmail = intent.getStringExtra(EMAIL) ?: ""
-
-        setContent{
+        setContent {
             ANDANDROIDTheme {
-                MyScreen(
-                    userName = userEmail
-                )
             }
         }
-    }
-
-    companion object {
-        private const val EMAIL = "email"
     }
 }
